+++ extensions/identity-hub-inmemory-store/src/main/java/org/eclipse/dataspaceconnector/identityhub/store/IdentityHubInMemoryStore.java
@@ -1,7 +1,6 @@
 package org.eclipse.dataspaceconnector.identityhub.store;
 
-import org.eclipse.dataspaceconnector.dtos.HubObject;
-import org.eclipse.dataspaceconnector.store.IdentityHubStore;
+import org.eclipse.dataspaceconnector.identityhub.dtos.HubObject;
 
 import java.util.ArrayList;
 import java.util.Collection;
