+++ extensions/identity-hub-inmemory-store/src/main/java/org/eclipse/dataspaceconnector/identityhub/store/IdentityHubInMemoryStoreExtension.java
@@ -3,7 +3,6 @@ package org.eclipse.dataspaceconnector.identityhub.store;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.eclipse.dataspaceconnector.store.IdentityHubStore;
 
 /**
  * ServiceExtension used to provide an in-memory store of HubObjects.
