+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/api/IdentityHubController.java
@@ -6,8 +6,13 @@ import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import org.eclipse.dataspaceconnector.dtos.*;
-import org.eclipse.dataspaceconnector.service.MessageProcessor;
-import org.eclipse.dataspaceconnector.service.MessageProcessorFactory;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageRequestObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.RequestObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.RequestStatus;
+import org.eclipse.dataspaceconnector.identityhub.dtos.ResponseObject;
+import org.eclipse.dataspaceconnector.identityhub.service.MessageProcessor;
+import org.eclipse.dataspaceconnector.identityhub.service.MessageProcessorFactory;
 
 import java.util.List;
 import java.util.stream.Collectors;
