+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/api/IdentityHubExtension.java
@@ -1,11 +1,11 @@
 package org.eclipse.dataspaceconnector.identityhub.api;
 
-import org.eclipse.dataspaceconnector.service.MessageProcessorFactory;
+import org.eclipse.dataspaceconnector.identityhub.service.MessageProcessorFactory;
 import org.eclipse.dataspaceconnector.spi.WebService;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.eclipse.dataspaceconnector.store.IdentityHubStore;
+import org.eclipse.dataspaceconnector.identityhub.store.IdentityHubStore;
 
 /**
  * EDC extension to boot the services used by the Identity Hub
