+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/dtos/HubObject.java
@@ -1,4 +1,4 @@
-package org.eclipse.dataspaceconnector.dtos;
+package org.eclipse.dataspaceconnector.identityhub.dtos;
 
 /**
  * data associated with MessageRequestObject.
