+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/dtos/RequestObject.java
@@ -1,4 +1,4 @@
-package org.eclipse.dataspaceconnector.dtos;
+package org.eclipse.dataspaceconnector.identityhub.dtos;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
