+++ extensions/identity-hub/src/test/java/org/eclipse/dataspaceconnector/identityhub/api/featuredetection/IdentityHubControllerTest.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.javafaker.Faker;
 import io.restassured.specification.RequestSpecification;
 import org.bouncycastle.util.encoders.Base64;
-import org.eclipse.dataspaceconnector.dtos.Descriptor;
-import org.eclipse.dataspaceconnector.dtos.MessageRequestObject;
-import org.eclipse.dataspaceconnector.dtos.RequestObject;
-import org.eclipse.dataspaceconnector.dtos.VerifiableCredential;
+import org.eclipse.dataspaceconnector.identityhub.dtos.Descriptor;
+import org.eclipse.dataspaceconnector.identityhub.dtos.MessageRequestObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.RequestObject;
+import org.eclipse.dataspaceconnector.identityhub.dtos.VerifiableCredential;
 import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
